requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.1.0,<2.3.0
openpyxl==3.1.2
supabase==2.8.0
google-generativeai==0.7.2
python-dotenv==1.0.0
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gunicorn==21.2.0
numpy>=2.1.1,<3.0.0
httpx>=0.25.0
postgrest>=0.13.0
sse-starlette==2.1.0
websockets>=12.0